- Moved 'diy' story template to 'doers-get-more-done'.
