+++ packages/stickers/src/doers-get-more-done/facebookIcon.js
