+++ packages/stickers/src/doers-get-more-done/instagramIcon.js
