+++ packages/stickers/src/doers-get-more-done/orangeCross.js
