+++ packages/stickers/src/doers-get-more-done/orangeDot.js
