+++ packages/stickers/src/doers-get-more-done/plus.js
