+++ packages/stickers/src/doers-get-more-done/rightArrow.js
