+++ packages/stickers/src/doers-get-more-done/youTubeIcon.js
