+++ packages/templates/src/raw/doers-get-more-done/index.js
@@ -22,26 +22,26 @@ import { __, _x } from '@web-stories-wp/i18n';
 /**
  * Internal dependencies
  */
-import { default as freshAndBrightTemplate } from './template';
+import { default as doerGetMoreDone } from './template';
 
 export default {
-  title: _x('Beauty', 'template name', 'web-stories'),
+  title: _x('DIY', 'template name', 'web-stories'),
   tags: [
-    _x('Health', 'template keyword', 'web-stories'),
-    _x('Bold', 'template keyword', 'web-stories'),
-    _x('Joy', 'template keyword', 'web-stories'),
+    _x('Doers', 'template keyword', 'web-stories'),
+    _x('Expand', 'template keyword', 'web-stories'),
+    _x('Start', 'template keyword', 'web-stories'),
   ],
   colors: [
-    { label: _x('Pink', 'color', 'web-stories'), color: '#f3d9e1' },
-    { label: _x('Green', 'color', 'web-stories'), color: '#d8ddcc' },
-    { label: _x('Black', 'color', 'web-stories'), color: '#28292b' },
+    { label: _x('Black', 'color', 'web-stories'), color: '#211f1e' },
+    { label: _x('Orange', 'color', 'web-stories'), color: '#ff630b' },
+    { label: _x('Cream', 'color', 'web-stories'), color: '#faf4ea' },
+    { label: _x('Light Grey', 'color', 'web-stories'), color: '#858280' },
     { label: _x('White', 'color', 'web-stories'), color: '#fff' },
-    { label: _x('Brown', 'color', 'web-stories'), color: '#eadfd6' },
   ],
   description: __(
-    'The modern and bright Beauty template lends itself well as a foundation for stories covering make up, beauty products, shopping guides, instructions & tutorials and more.',
+    'Motivate your audience to get out there and make something with the bold DIY template. Use it for DIY, crafting, 3D printing, woodworking or any other content targeting makers.',
     'web-stories'
   ),
-  pages: freshAndBrightTemplate.pages,
-  version: freshAndBrightTemplate.version,
+  pages: doerGetMoreDone.pages,
+  version: doerGetMoreDone.version,
 };
