+++ packages/templates/src/raw/doers-get-more-done/template.json
